Changing samplePeriod parameter type to float for StartTimedTelemetry.
Binaries submitted in 861652
Build time : 14 minutes 9.272000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 861647]

+++ BlueStatistics.cpp
@@ -201,7 +201,7 @@ void BlueStatistics::StartTelemetry( const std::string& server )
 	StartTimedTelemetry( server, 0 );
 }
 
-void BlueStatistics::StartTimedTelemetry( const std::string& server, int samplePeriod )
+void BlueStatistics::StartTimedTelemetry( const std::string& server, float samplePeriod )
 {
 #if CCP_TELEMETRY_ENABLED
 	if( s_isTelemetryConnected )
@@ -218,7 +218,7 @@ void BlueStatistics::StartTimedTelemetry( const std::string& server, int sampleP
 #endif
 }
 
-void BlueStatistics::StartTelemetryDump( const std::string& dumpFolder, int samplePeriod )
+void BlueStatistics::StartTelemetryDump( const std::string& dumpFolder, float samplePeriod )
 {
 #if CCP_TELEMETRY_ENABLED
 	if( s_isTelemetryConnected )
